Removing script for _Koalas to the Max dot Com_
I didn't really want to copy big chunks of source code, and the
developer has actually made it easy to use your own images.

The JavaScript function `btoa` can be used to encode the URL of the
image one wants to use:
    http://koalastothemax.com?<result>

But it's easier than that. Looking at `koala.js` (v1.3), the developer
(Vadim Ogievetsky) points out that one need not evaluate `btoa` and can
instead use the unencoded URL (as `<result>` in the URL pattern above).